Show welcome screen if opening of all projects was canceled or failed: IDEA-284273
If multiple projects are reopened, then showing welcome screen on cancel is disabled (see creation of OpenProjectTask inside RecentProjectsManagerBase#openMultiple, and its usage in ProjectManagerExImpl#doOpenAsync#handle). That is correct because we shouldn't show the welcome screen if opening of one project was canceled while the second one opened correctly.

Therefore, we perform an additional check after reopening multiple projects: if all of them failed, then show the welcome screen.

GitOrigin-RevId: aa2bd5184d28d12150a5391b29725e7fa79b3772