[documentation] optimization: don't reset the pane text unnecessarily
`DocumentationEditorPane` caches the text internally, so it's fast to access it.
However, when setting the text, it doesn't check if it's changed, so it resets the whole document.

This change also makes the scrolling position persist in the case:
- open doc in tool window;
- turn auto-update on (default);
- scroll in the doc;
- put caret on top of (or mouse hover over) another reference to the same element (class/method).

(cherry picked from commit 5efe785037c8eadc293c7eef7b592552134c4f9a)

IJ-CR-18398

GitOrigin-RevId: 6fd23ffaa26966449f023d2d8580e789c1b8e56c