ML in SE: Fix UnsupportedElementTypeException on tab switch (IDEA-284365)
The exception got thrown when SE tab was switched and results from unsupported tab (e.g. All) got reported from one that collects element features.
The fix checks the search start reason, and if it is a tab switch, it checks if all elements belong to the currently selected tab.

GitOrigin-RevId: 54a344b23b21d31b79e17b79a33142d5cc0ad310