Speed up DependencyGraphModelBuilder
Use only Node#id as the criterion for equals/hashCode, greatly speeding up the
redundancy checks during serialization. The ID was not unique in the whole build
before. I fixed that by reusing the graph generator for the whole build and having
an ever-increasing ID counter. The end-result is still the same. ReferenceNodes are
only created when a dependency is found multiple times in the same Configuration.

I've also removed the usage of Gradle's internal dependency report classes, which
were an unnecessary indirection. Instead, the graph generator now uses Gradle's
ResolutionResult API directly.

Fixes IDEA-283622

For the project mentioned in that issue, sync times are 2 times faster after this change.

GitOrigin-RevId: d8b03de6103d204deaef3c3b4876b0fa458ecb29