trusted-projects: fix situations when user answers Don't Open in openOrImport: IDEA-280859, IDEA-282352
openOrImport() returns a nullable Project, but the meaning of null not well-defined: it can be "we couldn't open the project via ProjectOpenProcessors and related logic, so you're free to open it some other way" or "project shouldn't be opened at all".

Most clients are OK with both meanings: they don't do anything extra on receiving null. But two of them behave differently: the CommandLineProcessor, and tryOpenFiles (invoked on drag-d-drop and on Open In -> IDE, see IDEA-282352). They treat null as a signal to try again, and it is really a valid case to try again (e.g. opening a txt file from the Finder context menu).

To fix this we introduce a special return value which is handled specially in the mentioned clients. Existing method is heavily used and is kept as is.

GitOrigin-RevId: 93ee094793c83ddfb5146b9ee7a6d66d63ae3c22