[vcs-log] provide methods to refresh log even when there are no visible tabs and power save mode is on
In power save mode (or when "vcs.log.keep.up.to.date" registry property is disabled) Vcs Log is initialized and refreshed only when at least one of the tabs is visible. If not, initialization is not performed and roots that need refresh are saved in the PostponableLogRefresher. This commit adds methods to check if Log is up-to-date and ensure it is by initializing it and performing postponed refresh.

GitOrigin-RevId: 32d79fc113ed93ab5c430cb0c67d9f2d8160ad73