RUBY-29087 Implement new RoR project generator UI
The new design is implemented only for Rails plugin, but not for RubyMine. In the future, it is also possible to add it to RubyMine and remove the outdated code that was left for compatibility reasons.

There are two problems with the new design.

The first problem with hints when validating fields. Validation itself works and the project is not created with incorrect input data. But the hints are not displayed above some fields. The corresponding issue is IDEA-282833.

Another problem is that additional Rails options should have been in the form of a drop-down list with options and tags. This component is non-standard IDEA-284304 and will probably require considerable time to adapt. Therefore, now Rails options are implemented as a text field with the ability to specify parameters just as if it were in the console. In the future, if there is time left, this text field will be replaced with the desired component.

IJ-CR-18732

GitOrigin-RevId: ac942e9a62209556992189e08c2d65e798b733cd