[kotlin] Quick fix to replace `T` type with `T & Any`
When a project/module uses language version 1.7, Java names with generic
types and `@NotNull` annotations are imported as having definitely
non-nullable types (e.g., `T & Any`).

This commit fixes the rendering of definitely non-nullable types in
quick fixes and their descriptions, and adds a quick fix bound to
TYPE_MISMATCH error to change function parameter types from `T` to
`T & Any`. It also prevents the unsafe cast of expressions of type
`T` to `T & Any`. The fix is not shown if `T & Any` syntax is not
allowed according to the module settings (`DefinitelyNonNullableTypes`
language feature is disabled).

More complex (flexible) types like `(Mutable)Collection<T & Any>` are
not yet supported as possible replacements (but are prevented from cast
as well).

^KTIJ-20425 Fixed

GitOrigin-RevId: 6d81e0c1fa29d0badf41bfb8883d09c638ee728c