Specify Kotlin LV and API Version explicitly if Kotlin facet is there.
This commit fixing the compilation in case of plugin 1.6.x version is installed.

By default, if we have some kotlin settings Kotlin facet appears and if no LV or API is specified
it uses the default values calculated from plugin version. So in case of 1.6.x IDE plugin it will be 1.6.
But we are using the ``-Xuse-old-backend` option that is not compatible with LV 1.6

(cherry picked from commit 212c99f69e1c7358657272ab9ab7fd4112263ff2)

GitOrigin-RevId: 8e559e18208df2c02ef230a063bb4038c10278b9