[javadoc] IDEA-241299 Support the @index Standard Doclet tag
Revert the changes in lexer that were introduced after adding the index tag. In order not to complicate the lexer and parser the index tag doesn't handle its value in a special way regardless if its value is in quotes or not. The correct generation of the javadoc is handled inside JavaDocInfoGenerator#generateIndexValue now

GitOrigin-RevId: 78643f7b199aa1592e95c54ffbe59983babfb596