[vcs] IDEA-284854: Disable Commit action during VCS initialization if non-modal commit is used
It is still possible to open commit dialog when VCS is initialized but workflow is not. We can add check for that and disable Commit action if workflow is not initialized, but it may cause unpredictable disabled Commit action (e.g. it will be disabled in steps of this issue IDEA-284764)

GitOrigin-RevId: 2dd0e3c1de8020b1655e66fba2fb23d6f1eca5a8