deprecate `ModalityInvokator`
The returned callback might be rejected if expired in `invokeLaterWithCallback#invokeLaterWithCallback`,
and at the same time the callback might be done if expired in `FlushQueue#pollNextEvent`,
which basically says that it's incorrect.

It's too late to change anything in its implementation, but the good thing is that this interface had exactly 1 usage in
our sources. There is a couple of usages in external plugins which don't even need the returned callback, and a couple
of usages in the deprecated code in MPS.

GitOrigin-RevId: d6dc13d050653e657e95f849c45c9831ffc0c112