settings-sync: implement resolving a conflicting merge
The simplest strategy is implemented for now: if Git fails to resolve conflict using the recursive strategy, then we use either YOURS or THEIRS strategy. The side to choose is defined by the latest commit made to either 'ide' or 'cloud' branches.

If there were two conflicting files, one of which was modified later in 'ide', and another was modified later in 'cloud', then this strategy will resolve incorrectly, since it would prefer a single branch for both files. This case seems to be rare and requires a more complicated implementation, so it was decided to skip it for now and possibly implement in the future.

GitOrigin-RevId: 44c540a65d9df13241dc677c61a3b30d75af63ae