Remove legacy LenientConfiguration API usage
This API is outdated and inefficient, adding overhead to the sync process that shouldn't
be necessary. The only information that the ArtifactView API doesn't provide is the concept
of classifiers and packaging. These are Maven/Ivy implementation details Gradle deliberately
no longer exposes. Instead, there is the new ResolvedVariant API that IntelliJ could use if
you wanted to provide more detail about which variant of a library was resolved.

This change is more of a basis for discussion than being intended to be merged as-is. I think we
should finally let the old LenientResolutionResult API die and the first step would be IntelliJ
no longer using it on modern Gradle versions.