[resource bundle] IDEA-285267 Text overrides existing entries.
When a new property was added to a resource bundle the code used to invoke StructureViewComponent#select which returns a promise and its code didn't get executed until the current method exited which made a mess with updating editors on the resource bundle editor panel. Now the code to select the newly added property is pulled from the promise back to the synchronous part and it correctly reinitializes editors.

GitOrigin-RevId: 2da9adf7b2cae090c31aa1e7df76f215e7bd84b8