IDEA-285713 Fix partially showing headers on Linux
Thanks a lot to Anton Tarasov for the help to figure this out, and
for providing the patch.

The issue was due to a lack of scaling which was biting us only on
Linux with a non-integer UI scaling factor. On Linux, that works in
the legacy "IDE-managed HiDPI mode". Other OSes don't have this
limitation for float scaling factors, they work in JDE-managed mode.

GitOrigin-RevId: 1ef65dfb73904a7174babfe10b2b23dcf3f14e70