Show the welcome screen if opening the project on start fails: IDEA-286542
For instance, if the user drops a file to the application icon, and then resists opening the folder in the Trust Project dialog, then the project instance is null. If there are no projects to reopen, the previous code used to stuck on the splash screen. It should show the welcome screen instead.

GitOrigin-RevId: d05e4492886fe552353f07d3d3c5d2f36a0d2ec1