settings-sync: remove the ApplicationLoadListener
It is not necessary to initialize the Setting Sync at such an early loading stage. It is also inconvenient because the service manager is not available at that point, so we can't check if the user has enabled the sync or if they have disabled it.

Instead, don't initialize the service until the first synchronization request which, currently, will happen on the initial application frame focus gain.

GitOrigin-RevId: 2061012d10e3e6f87a13aeb47ab8b5db4151faab