PY-52575 Attach Python tools to "intellij.python" in the licenses list
Because "intellij.python.helpers" is not included in the distribution by
the standard means, we ended up not including the information about libraries
attached to it on the page of third-party software used in PyCharm.

The only exception was MathJax library that was attached to "intellij.python".

(cherry picked from commit aea307e82ef57e9424f555bf06e620ffb64c8a24)

IJ-CR-19854

GitOrigin-RevId: ca8b97141983649a96edfd434ddd572645a6ef60