KotlinBuildProcessParametersProvider: Fix exception on project open
^KTIJ-20814 Fixed

Don't throw exception when kotlin home path doesn't exist yet.
Unfortunately, `KotlinBuildProcessParametersProvider.getPathParameters`
can be called during "is up-to-date" check in JPS.

It looks strange for me that JPS launches build process during
"is up-to-date" check. But it seems that JPS launches build process
in some special mode and doesn't really do the compilation so this
fix should be safe.

GitOrigin-RevId: 374fa53ef62f1d09e576429b30c4ea88b45f395a