Improve linkProjectLibrary performance IDEA-286459
GradleProjectResolverUtil.linkProjectLibrary is invoked at least once
for each library dependency in each module by either Gradle or Android
plugin and it each time it searches for an existing library data
matching the one being added by name.

Assuming LIBRARY_DATA nodes are only added to the collection and the
collection is never cleared, it is safe to cache this mapping and avoid
expensive linear search on each invocation.

This change adds a cache as a user data attached to PROJECT_DATA data
node.

This can be further improved, if we assume that LIBRARY_DATA nodes
should never be added in any other way than via linkProjectLibrary. In
this case we can avoid linear searching even when adding the library for
the first time.

GitOrigin-RevId: 5fd95b64547194ad9d6270eb0ad58bbe4299e60e