RUBY-29252 added tests illustrating the problem and found the cause
The problem is missing headers in ps output. Unsure why. But works without it (at least list is not empty).
Current tests checks only ps/lsof parsing logic and generic actions building.
It does not cover matching process path with sdk path, we will do this separately if and when necessary.

GitOrigin-RevId: 08415a09e5eccfe0d26cb4e5554b6484bdc19292