[git-stash] do not keep full stash commit object in memory
Stash commit contains changes to all parents, including the commit with untracked files, if they are saved in the stash. As the untracked files commit has only untracked files, and not other repository files, the number of changes with the stash commit could be potentially huge and should not be saved in the cache.

GitOrigin-RevId: c63971074979dd74a61405c9dd73dd31f3b3a5d5