[Workspace Model] IDEA-265327 Vanish artifacts duplicates on project loading
Along with modules and libraries, artifacts started to report duplicates where two xml files with ArtifactName.xml and ArtifactName2.xml are created.
With the current reports it's not fully clear how this situation can happen, so additional information was added to the error message.
Also, logic similar to libraries about removing duplicates, was added.

GitOrigin-RevId: 051e9dc72fc376586d95374a919183c554c9212c