[kotlin] Allow facade with forced platform for JVM-modules
Allow creating facade with forced platform even if the current module is
JVM.

It was assumed that in such cases forced paltform isn't needed, as the
resolution will use refiniment as needed. That's indeed the case, but
the issue is that evaluator has more elaborate requirements here, as it
will pass resolution results in form of BindingContext into backend (to,
well, generate the bytecode for evaluated expression)

Turns out, not all records in BindingContext are properly refined, and
some of them (for example, RESOLVED_CALL-slice) contains references to
the common descriptors. Actually, it was an explicit architectural
choice for type refinement: that only a very few placesare refined
proactively; for all the rest usages clients need to call refinement
themselves.

It doesn't seem feasible to teach backend and evaluator to refine all
information, accessed from the BindingContext, manually.

This means that the BindingContext for evaluator should always be built
as if we were analyzing common and JVM sources together. I.e., we should
use forced facade always for MPP projects, even in JVM-specific code.

This commit just lifts this limitation

^KTIJ-20775

(cherry picked from commit 35de7c2acd161e2ddb11c3766e3b4909187746be)

GitOrigin-RevId: ac1fc41e8e092077457f9d154078a5316a2d5283