[kotlin] Allow facade with forced platform for special files
Initially, it was assumed that facade with forced platform will never be
needed for "special files". This is not correct, because all files out
of project's source roots are considered "special". This specifically
applies to sources of project' libraries.

Therefore, it's a normal situation when one wants to request facade with
forced platform while some files are special (suppose, we're standing
with JVM evaluator in sources of a common part of a library)

However, that notably requires to wire the PlatformAnalysisSettings into
various special-files-related caches. Due to how the current code is
structured, some codepaths also changed, while actually they don't need
that change (notably, scripts)

GitOrigin-RevId: 5fc2ea486f52c6252906895dd7be975d55a3840d