[groovy] IDEA-287303: Reuse results of ResolveCache in PartialContext
Before this change, if inference entered into PartialContext, it resolved __all__ references it needed. This had terrible effect on the performance. It is still impossible to publish globally the results of PartialContext, but as it is known that global cache results are correct, it wouldn't do any harm to use them in non-cacheable environment

GitOrigin-RevId: f781774a1ff8a97a41a362d30b84edcf806136e9