Additional changes from [import] Fix Kotlin facet creation with convention plugin
In kt-212-master we have the same commit, but a bit different version
of it: 277c81a43bd91f16a and this version was added into kt-213-*
branches.

In 213 -- we have "the full version" under aedcfdddef9b544fddee6f6b.

So this commit just adds the missing changes from 213 branch

GitOrigin-RevId: 4e621d0663e69ff7ea098d0928df64ffbbf77e31