[git-stash] show a warning when index contains files other than selected
When the "git stash push" command is called for the list of files, the whole index state is saved to the stash regardless of which files were specified to the command. Because of this, stash created in this situation would contain unexpected changes. The warning is meant to notify the user about such behavior.

GitOrigin-RevId: 92ae7335703096a62aa06d89f09ccb63e84e8af7