RUBY-29279 fix parsing of number literals with underscores.
Due to parsing optimizations introduced in RUBY-26170, when underscores after numbers were incorrectly parsed as identifiers, they stayed as such until a full-blown reparse was triggered. Fixed the optimization logic, so that full-blown reparsing is now happening when we try to reparse an identifier that starts with an underscore and is right after a number literal. Added test cases for different types of integers (decimal, hex, octal, ...).

Check condition was modified to also take into account the current identifier's text. We check that current text starts with "_", before proceeding with more expensive checks. We also make sure that when we remove extra underscores ("100__0" --> "100_0"), we trigger a reparse. Added a separate test for this edge case.

Review IJ-CR-20390.

GitOrigin-RevId: a0692451fdb366867bc5ee41ea646764e0bc1e61