[kotlin] Prevent "Change type" fix for 'T & Any' if DNN feature is off
Using `-Xenhance-type-parameter-types-to-def-not-null` compiler
argument, it is possible to enable Java `@NotNull` type enhancement to
definitely non-nullable types while the language feature
`DefinitelyNonNullableTypes` is still disabled (e.g., when the language
version is set to 1.6).

In this case, the "Change parameter type"/"Change type" quick fixes do
nothing instead of changing the type `T` to `T & Any`, as the `T & Any`
type is resolved not as a DNN type but as a regular type parameter `T`.

This commit adds an explicit check for the state of
`LanguageFeature.DefinitelyNonNullableTypes`, so the useless fix is
not proposed to the user.

^KTIJ-20981

GitOrigin-RevId: a249bdd00042261904aaf38c95850093079ee5ce