[git-stash, shelf] include stashes and shelves with matching files in speed search results
Do not trigger or wait for loading changes in stashes or shelves, but use cached changes assuming that they are always preloaded in background. Use only file names and not paths for now. If a stash or shelf has matching files, highlight the whole corresponding node in the tree with the same style as normal matches. Also highlight matches in the changes browser by providing it the special "combined" SpeedSearchSupply instance which is aware of both text speed search in the changes browser and files speed search in the stashes/shelves tree.

IDEA-288154

GitOrigin-RevId: 9563fcda00b7dd082514c8477676a7a78b25266e