[Workspace Model] [IDEA-288277] Do not try to save artifacts with conflicting names
At the moment, when the user renames artifact, IJ saves the new name on every keystroke. That means, that during typing, one artifact may have the same name as some other artifact even if the target name is not the same.
Workspace model doesn't allow us to store artifacts with the same names, so we should not update the artifact name in case it already exists.

GitOrigin-RevId: a0565388e618dbbcbd6524cbf22e54c4748983f0