PY-52868 Don't compare version strings lexicographically
Because of that versions like 3.11 were considered lower than 3.9.
Then we considered 3.9 to be the latest version configured for the inspection
and, thus, suggested enabling 3.10 and 3.11 in it over and over again.

Also, I added exclusion of duplicate versions when enabling new ones in
the inspection settings, just in case.

GitOrigin-RevId: 69f40e7bc1b95c564beeadf466b64f9f653feec6