[ExpUI] Fix default checkbox and radio bg in Dark theme
Darcula has same background in checkboxes for default and selected states. In the new Dark theme, checkboxes have different backgrounds in default and selected states. Added checkbox and radio SVGs for default state with correct colors. SVG files have 'Temp' in the name: to be removed in 2022.2 when we have correct SVG files for all checkbox and radio states.

GitOrigin-RevId: 48cce508a6a6fcb1417a40cc3779cff337b93739