[kotlin] Fixed infinite recursion in `MakeUpperBoundNonNullableFix`
There was an infinite loop in the quick fix factory for "Make upper
bound non-nullable" quick fix during the search for type parameters that
would differ in nullability. The loop occurred when processing some
types containing the star projection (e.g., `Enum<*>` and
`Class<Enum<*>>`).

This commit fixes the infinite loop by adding an
explicit check to skip the recursive call if any involved type
projection is a star projection.

^KTIJ-21154 Fixed

GitOrigin-RevId: 334c4ea9d094ca4974d300f78544feb8460a81a7