[kotlin] Fixed infinite recursion in `MakeUpperBoundNonNullableFix`
There was an infinite loop in the quick fix factory for "Make upper
bound non-nullable" quick fix during the search for type parameters that
would differ in nullability. The loop occurred when processing some
types containing the star projection (e.g., `Enum<*>` and
`Class<Enum<*>>`).

This commit fixes the infinite loop by adding an
explicit check to skip the recursive call if any involved type
projection is a star projection.

^KTIJ-21154 Fixed

GitOrigin-RevId: 3f8b886d8148fb609f9cc0669c1cf7f3754bff08