[kotlin] Use class from kotlin-compiler-common-for-ide for dist artifact resolution
As there are two kotlin-stdlib libraries in the project (kotlinc.kotlin-stdlib and kotlin-stdlib-jdk8), the wrong version might appear in classpath first. It may be especially harmful for local development in kt-branches, as -for-ide artifacts are located in a local Maven repository, while kotlin-stdlib-jdk8 is likely to be inside ~/.m2/repository.

GitOrigin-RevId: 66f0f355e36b3e12c7f3454220d9558056eb4603