[kotlin] Use class from kotlin-compiler-common-for-ide for dist artifact resolution
As there are two kotlin-stdlib libraries in the project (kotlinc.kotlin-stdlib and kotlin-stdlib-jdk8), the wrong version might appear in classpath first. It may be especially harmful for local development in kt-branches, as -for-ide artifacts are located in a local Maven repository, while kotlin-stdlib-jdk8 is likely to be inside ~/.m2/repository.

GitOrigin-RevId: 24f8dc353ca8949536f2b5540e690b1c0b9520c3