settings-sync: synchronize access to xmls between the stream provider and the process updating from cloud: IDEA-287347
This doesn't solve the simultaneous update both from the IDE and from the cloud: there is still a possibility that cloud settings could overwrite local ones if performed at the same time. This change synchronizes the file access, so that there are no file system exceptions both from the storage subsystem and the settings sync.

GitOrigin-RevId: 8a3daa8e1cabe593179dc35ffd01d4f8d3c742ec