IDEA-289082 execution context switches
This change adds a way to switch from a coroutine to the blocking context
(thread local `currentThreadContext` with job/sink/etc): `com.intellij.openapi.progress#blockingContext`.

A global indicator and a current job could exist simultaneously in the current thread, e.g. when a client explicitly
uses `ProgressManager` APIs to run something under a `ProgressIndicator` in a thread, which already has an associated
job via cancellation propagation IDEA-276966. In such case should prefer the indicator over the job to maintain
the compatibility. But when doing an explicit switch with the newer APIs, the block of code is being run
either under a job (via `currentThreadContext`) or in a coroutine, so no indicator APIs are expected to work there.

This change also fixes IDEA-286739:
```
withIndicator { // root indicator
  runBlockingCancellable { // 'child' of indicator
    readAction { // child of runBlockingCancellable
      // current job available here
      runBlockingCancellable {
        // should use the current job, but indicator was preferred before the change
      }
    }
  }
}
```

If `runBlockingCancellable` was modified to prefer the current job, then IDEA-286739 would also be fixed,
but the following would work incorrectly:
```
withJob {
  // current job
  runBlockingCancellable { // child of current job
    runUnderIndicator { // switch from coroutine to indicator
      runBlockingCancellable {
        // would become a child of outer current job,
        // should be a child of the inner-most indicator
      }
    }
  }
}
```

GitOrigin-RevId: 33cf34c69dbdf451f29c1fb1be98ca22bc83e1cf