[vfs] IDEA-279039  Fix VFS freeze by checking existence of archive based on `VirtualFile` first
Previously `ArchiveFileSystem.getAttributes` checked the existence of the archive by calling
IO operation to disk and on the Windows system it was slow. The idea was to check the existence
at LocalFileSystem. But we can't simply rewrite the old method because it used in many cases
e.g `RefreshWorker.scan` and base on the existence or not archive on disk `VFileDeleteEvent`
is fired(and record will not be removed from VFS and PresistenseFS). Thus, it was decided to
introduce new method where checking existence on disk is absent.

GitOrigin-RevId: af5a2c0150e7377c6498a521ec6da313fe4541e8