settings-sync: cleanup: use EventListener instead of MessageBus for internal settingsChanged events
With all its advantages, the MessageBus has a downside since the event delivery can fail due to an unrelated failure happened while delivering another event in the queue. Delivering settings events is critical, so a more reliable method is to be preferred. Since the whole event communication is internal for the settings-sync plugin, it is easy to use a simple EventListener.

GitOrigin-RevId: 57346014d18d343cfe514123d089cca93648084c