ClasspathBootstrap: modify checking for BANNED_JAR to avoid getting IDEA home path.
The project we build may not be intellij project at all and this is the sources of JPS compiler.
For now, modifying this check to prevent getting IDEA home path if jar names is not app.jar

In the future it would be best to remove intellij monorepo-specific code from compiler sources.

GitOrigin-RevId: 92214dbf2947caa842be12f4cfb1ec20cdcd7eef