[IFT] Fix debug actions highlighting
Found that there are two action toolbars in the same place that are showing, but one of them have height = -1. So UI search finds this toolbar but can not highlight it.
With this change 'showing' components with empty bounds will not be accepted by UI search. Also, if bounds of some component becomes empty, the rehighlighting will start.

IJ-CR-21162

GitOrigin-RevId: 931fea5187ab593275ac0028235d998b596383dd