DS-3209 Fix improper change of request in PySciRestService.isSupported(), which cut off query part of request URI
The same mutable `request` object is passed to `isSupported` method of every registered `HttpRequestHandler` until the first handler that accepts the request. `PySciRestService.isSupported` cut off query part of request URI in `request`. If another handler `JupyterCefHttpHandlerBase` was in the chain after `PySciRestService` this change of the URI failed the rendering of Jupyter cells.

GitOrigin-RevId: fcc57022a76c1ed5b7ea44ab38ae59f0f9c3b845