settings-sync: improve initialization
Initialize the plugin "somewhen" after the app startup, and not by the first frame activation. The "somewhen after startup" is currently defined by componentsInitialized.

On startup we initialize the SettingsSyncPluginManager, and begin listening to the "Enable Settings Sync" button in the settings. If Settings Sync is already enabled in the settings, we schedule plugin initialization on a pooled thread, otherwise we do nothing.

GitOrigin-RevId: 0e9c2fb85c1a957b4eec362cb4ab018be164fd83