[fleet] Backend: fixed race condition when marking files as dirty
The test RenameTest::`rename properly updates references` was flaky.
The reason was that FSDFileSystem was only refreshing the file system
after explicit call from the FileWatcher. During this update, the entire
list of dirty files is consumed and reset via FileWatcher::getDirtyFiles,
which is a mutable operation. However, There is another entry for
FileWatcher::getDirtyFiles and that is the `markSuspiciousFilesDirty` method
which is used for manually watched files. In this entry point were not
properly refreshing, hence missing dirty file updates.

Note tha this was initially fixed in 7c77fb8b28db01d2f41337f7a07a11ab31d860de,
but the change was revered in 9cd6593903da6a0bfad5f97aff12e6a4b745c4d5.
The issue is that markSuspiciousFilesDirty is eventually called during the
refresh process, and if we trigger a refresh there, then we end up in an
endless loop.

To solve the endless loop problem, we only trigger refresh in
markSuspiciousFilesDirty if there are regular dirty files marked, but
not if there are manually watched roots (aka suspicious), because this would
trigger and endless loop.

This means that we will not properly trigger refresh for manually watched
roots, but given that FSDFileSystem is the only user of this refresh
trigger and we don't seem to use manually watched roots here, it seems ok
to me.

Related: FL-5172

GitOrigin-RevId: cd9b0c72e5b761c3af1fb1f09349b8e33f2d1fad