[kotlin] Filter out synthetic enum methods in dumb light classes
Synthetic enum methods are no more filtered in CLS stubs. It seems to be a cleaner and more correct approach, as enum methods are available in all present instances of 'PsiClass'. However, it requires to filter methods on stub usage in some cases. In particular, 'ownMethods()' should only return methods present in source code.

GitOrigin-RevId: 50933d7d4cada4aea152a882256c5d4e3afacbe9