[workspace model] improve performance of applying changes to facets (IDEA-290047)
'prepareForCommit' is called for each module with facets, so if we process facets from all modules in it, the execution time is quadratically dependent on number of modules. Now the total time is linearly dependent on number of facets.

(cherry picked from commit c32ebbf3d51fff4eec2cfac39a296be7beac3573)

Change-Id: I73b97094538a3270d41c2a95fb68504a8d747394