[workspace model] improve performance of applying changes to facets (IDEA-290047)
'prepareForCommit' is called for each module with facets, so if we process facets from all modules in it, the execution time is quadratically dependent on number of modules. Now the total time is linearly dependent on number of facets.

(cherry picked from commit 6fd729b7624f23b9509fd16d947657ae45b00904)

IJ-CR-21695

GitOrigin-RevId: 577b4bde0f8daf72c7e0e9adb2c8931daf68176e