1/2 [refactoring] Make it possible to reference projectFilePath when PathMacroManagers are created
This commit is just a refactoring (change in code which doesn't change
any behaviour) commit. This refactoring commit is required for the next
commit.

Point of this commit is to make it possible to reference
`configLocation::projectFilePath` (this method will be created in the
next commit) or `project::getProjectFilePath` in all places where
`ModulePathMacroManager.createInstance` or
`ProjectPathMacroManager.createInstance` are used.

GitOrigin-RevId: f8cc66f9877b89eec87067e94a35797e1c7d70f8