IJ-MR-21596 1/2 [refactoring] Make it possible to reference projectFilePath when PathMacroManagers are created
This commit is just a refactoring (change in code which doesn't change
any behaviour) commit. This refactoring commit is required for the next
commit.

Point of this commit is to make it possible to reference
`configLocation::projectFilePath` (this method will be created in the
next commit) or `project::getProjectFilePath` in all places where
`ModulePathMacroManager.createInstance` or
`ProjectPathMacroManager.createInstance` are used.

(cherry picked from commit 1a1282903ed58849d64ae46a83b08ab15e6b0dd5)

GitOrigin-RevId: 58fbde89bd93de119de5498726255e9fe0e61e5d