RUBY-28322 fix problem with bundler installation before installing missing gems.
Problem was in accidentally inverted condition, due to which we offered to install 'bundler' gem when it was already present and didn't offer when it was missing.

Fixed by inverting the condition and adding an integration test to cover this scenario.

Review IJ-CR-22063.

(cherry picked from commit 59f301bcb4ded8e82fb7fcde2f923823a453b5f6)

GitOrigin-RevId: b2ed942e1fe027bf1ae200529adeb94f59c81457