PY-52953 Fallback to the first module of a project to find SDK for a scratch file
For scratch files, ModuleUtilCore.findModuleForPsiElement() returns null, hence
we fail to detect the corresponding project SDK and have to use any configured
SDK to launch pycodestyle.py. If it was <3.10, unlike the real project SDK, it
triggered false positives about the pattern matching syntax.

At the moment, there are several approaches to finding SDK for a non-module
file: PyBuiltinCache.findSdkForNonModuleFile(), PyOpenPStatAction.findPythonSdk()
and DocStringUtil.getModuleForElement(). All of them should probably be somehow
unified in the future.

GitOrigin-RevId: 80378d52da54503f2462cd800d17ccc184629df9