[kotlin fir] fix ISE Cannot find entity for library with ID LibraryId(name=myLibrary...
This was caused by the fact that map inside ProjectStructureProviderIdeImpl was storing references to outdated libraries as keys.
Even those keys was never queried directly, equals was still called on them to check for collision in hash map.
And on invalidated library that caused the exception.

The presence of ModuleInfo -> KtModule cache is questionable, so it's removed for now.

GitOrigin-RevId: 6ce853ba9db3ef3dd2b38856538f9a0bbff4b51a