KTIJ-1236 Ignored test has a run test gutter icon
The initial idea of having run icon for ignored test is to allow
its execution in individual mode (differ to class/package/module).
It simplifies understanding ignore reason and further fix.

Unfortunately, we cannot control test execution for all build systems
and test frameworks. Gradle in particular has its own built-in support
for JUnit but doesn't allow fine-tuning behaviour. As of now launching
ignored tests is impossible.

With this commit we remove run-icons support for ignored tests in
Gradle modules.

^KTIJ-1236 fixed

GitOrigin-RevId: 7ac88e9c920033df5d47cd5cd676a9b7dbeb04b4