[kt-213-* only] Fix compilation via JPS in debug IDEA
Basically this code is the copy from the `kt-221-master` branch.
But in the `kt-221-master` there is additional logic about
unbundled JPS, but we don't have such logic here, because
in 213 IDEA we have only bundled JPS.

About initial problem. `compileServer.plugin` way to set up
even before this commit was useless, because there is 2 cases:
1) IDEA running from sources. In this case it isn't working, because
there are no modules with the name like `kotlin.jps-plugin`.
That is because jps sources were moved back to Kotlin repo
2) production IDEA -- then there are no files like
"jps/kotlin-jps-plugin.jar;kotlin-jps-common.jar;kotlin-common.jar"
currently in the plugin. jps is just missing, and kotlin-jps-common
is actually called kotlin-jps-common-for-ide-$version$.jar

Next step is to fix the resulting plugin layout, so the production
IDEA with JPS plugin will also work.

GitOrigin-RevId: 2aef999d761990f26b69fd302ed0f09fb2e8dffc