IDEA-291451 Prevent deleted root module from being restored
This is done by temporary disabling the fix for IDEA-288703 in IDEA (but not in minor IDEs)

The resulting behaviour currently is not actually correct (but probably good enough until the bugfix update), and here's an example to demonstrate why:
• Create an empty project
• Delete the only module
• Close the project
• Reopen project
• For a brief moment you'll see the recreated root module, but then it disappears
What happens is that the project is initially loaded using `WorkspaceModelCache`, and a default root module gets created by `PlatformProjectOpenProcessor` (see `ProjectManagerExImpl#prepareProject` for more details). This new project state (with the newly created root module), however, does not get serialized, and therefore, once the project is fully loaded, `DelayedProjectSynchronizer` overwrites this state with the module-less state that was serialized before the project was closed after deleting the module.

This behaviour is, obviously, flawed. But it's been like that for god knows how long – a couple years probably, since it was probably introduced once `WorkspaceModelCache` was implemented. The commits addressing IDEA-288703 did also fix this weird stunt, but this commit disables that fix for IDEA. Proper comprehensive solution is to be shipped in 2022.1.1 bugfix update.

(cherry picked from commit 430a8425115ccb9264aac3ee0fd746500a1aa52e)

IJ-CR-22998

GitOrigin-RevId: f48c0fbc73ff4ada58e64cfb333c0ab95bbf4933