Extend Selection: prioritize blank lines above
If a whole method or class is selected, make Extend Selection select
the blank line above it if possible, instead of below. This allows to
cut the selection to the clipboard and then paste it somewhere else
by putting the cursor exactly between two methods or classes.
With the previous behavior, it was awkward: pasting between methods
or classes would lead to two consecutive blank lines after the pasted
method/class. To get the desired result, one would need to place
the cursor exactly one line below the blank line separating methods
or classes, which was a minor annoyance.