Jump to type source by Ctrl+click on a debug inlay
Jump to Type Source is a very convenient debugger action when the type
at runtime differs from the compile-time type, or when a variable declared
as val/var without explicit type, so there is no type in the source to
Ctrl+click. However, it’s somewhat awkward to access: one has to either
access a context menu or select the appropriate value in the watch list
and press Shift+F4.

This commit introduces a more convenient way to navigate to the source
of the runtime type of a value by simply Ctrl-clicking the appropriate
debugger inlay in the editor.