IDEA-271068 vcs: fix 'Before Commit' options synchronization between CommitOptions and Configurable
Problem: a commit option might be changed via Configurable.
This change will not be noticed by an already created RefreshableOnComponent.
This might lead to a consequent 'saveState()' call reverting changes made via Configurable.

This is not an issue for modal commit dialog, as it blocks showing Settings dialog by itself.
For non-modal commit mode we need to refresh RefreshableOnComponent immediately before using and
save changes as soon as options popup is closed.

This change breaks previous behavior of "restoreState() and saveState() are called
once for each RefreshableOnComponent (right after <init> and right before the actual commit)".
This behavior was not documented, and seemingly was not exploited by existing custom plugins.

GitOrigin-RevId: 1b15cb1cac7349fe9f962ffa1d8be916a2809595