WI-66304: extract subBlocks for all BladeRootBlocks after building template blocks
By default, buildTemplateLanguageBlocksInside() method extracts subBlocks only for a root composite block. In case when file starts with a comment, BladeRootBlock is not a root, because the comment is out of the BladeRootBlock. In such case, a new composite block that contains the comment and the original BladeRootBlock is created, and only subBlocks of this block will be extracted by default.

(cherry picked from commit c9248886db01cf9eea73598caea8c16c534a65c6)

IJ-CR-23880

GitOrigin-RevId: 12e69d9f81ba89ae60cb36dce006dbda61b8e71d