IDEA-290974 initialize UI after the content is loaded and make sure the UI actually uses the loaded content
The UI is instantiated _after_ the content is loaded, and then the UI is shown
immediately in the same EDT event in `#showDocumentationPopup`.

The UI instantiation uses undispatched coroutines (they are executed until the first suspension point in the EDT),
which handle non-empty `DocumentationBrowser#pageFlow` (by design) with non-empty `DocumentationPage#contentFlow`
(it's non-empty because this happens _after_ the content is loaded), which actually initializes the UI editor pane.

GitOrigin-RevId: a5c61ae552b25b30eae07040bf4b41cabd7ee0e0