[EE] Force Evaluator tests to run in Unspecified JVM platform
The Kotlin expression evaluator runs _on top_ of the loaded project. Hence, it needs to reuse the resolution facade for both IDE compilation and during analysis of code fragment for compilation.

The evaluator is forced to use `unspecifiedJvmPlatform` to improve the user experience of the evaluator.

This also happened to align with all evalautor test suites until now.

The MPP tests used a different platform, hence got a different resolution facade, so the analysis was "fresh" and descriptors was not reused across project and fragment code leading to unbound references in psi2ir.